directory restructured and fixed dependency
