--- datasetinsights/data/simulation/__init__.py
@@ -1,12 +1,0 @@
-from .captures import Captures
-from .metrics import Metrics
-from .references import AnnotationDefinitions, Egos, MetricDefinitions, Sensors
-
-__all__ = [
-    AnnotationDefinitions,
-    Captures,
-    Egos,
-    Metrics,
-    MetricDefinitions,
-    Sensors,
-]
