+++ datasetinsights/estimators/faster_rcnn.py
@@ -14,8 +14,8 @@ from codetiming import Timer
 
 import datasetinsights.constants as const
 from datasetinsights.data.bbox import BBox2D
-from datasetinsights.data.datasets import Dataset
 from datasetinsights.data.transforms import Compose
+from datasetinsights.datasets import Dataset
 from datasetinsights.evaluation_metrics.base import EvaluationMetric
 from datasetinsights.torch_distributed import get_world_size
