+++ datasetinsights/stats/visualization/object_detection.py
@@ -3,8 +3,8 @@ import dash_html_components as html
 import dash_table
 import pandas as pd
 
-import datasetinsights.data.simulation as sim
-import datasetinsights.visualization.constants as constants
+import datasetinsights.datasets.simulation as sim
+import datasetinsights.stats.visualization.constants as constants
 
 from .plots import histogram_plot, rotation_plot
