+++ tests/datasets/test_statistics.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from datasetinsights.data.datasets.statistics import RenderedObjectInfo
+from datasetinsights.stats.statistics import RenderedObjectInfo
 
 
 def test_read_filtered_metrics():
