+++ tests/datasets/test_synthetic.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 import pandas as pd
 
 from datasetinsights.data.bbox import BBox2D
-from datasetinsights.data.datasets.synthetic import (
+from datasetinsights.datasets.synthetic import (
     SYNTHETIC_LOCAL_PATH,
     SynDetection2D,
     _get_split,
@@ -14,7 +14,7 @@ from datasetinsights.data.datasets.synthetic import (
 )
 
 
-@patch("datasetinsights.data.datasets.synthetic._download_captures")
+@patch("datasetinsights.datasets.synthetic._download_captures")
 def test_syn_detection_2d(mock_data):
     parent_dir = Path(__file__).parent.parent.absolute()
     mock_data_dir = str(parent_dir / "mock_data" / "simrun")
