+++ tests/simulation/test_download.py
@@ -13,13 +13,12 @@ from datasetinsights.data.download import (
     download_file,
     validate_checksum,
 )
-from datasetinsights.data.exceptions import ChecksumError
-from datasetinsights.data.simulation.download import (
+from datasetinsights.data.exceptions import ChecksumError, DownloadError
+from datasetinsights.datasets.simulation import (
     Downloader,
-    DownloadError,
+    FileType,
     _filter_unsuccessful_attempts,
 )
-from datasetinsights.data.simulation.tables import FileType
 
 
 @pytest.fixture
@@ -60,7 +59,7 @@ def test_download_bad_request():
 def test_download_rows(downloader):
     n_rows = len(downloader.manifest)
     with patch(
-        "datasetinsights.data.simulation.download.download_file"
+        "datasetinsights.datasets.simulation.download.download_file"
     ) as mocked_dl:
         matched_rows = pd.Series(np.zeros(n_rows).astype(bool))
         downloaded = downloader._download_rows(matched_rows)
@@ -68,7 +67,7 @@ def test_download_rows(downloader):
         mocked_dl.assert_not_called()
 
     with patch(
-        "datasetinsights.data.simulation.download.download_file"
+        "datasetinsights.datasets.simulation.download.download_file"
     ) as mocked_dl:
         matched_rows = pd.Series(np.ones(n_rows).astype(bool))
         downloaded = downloader._download_rows(matched_rows)
@@ -79,7 +78,7 @@ def test_download_rows(downloader):
 def test_download_all(downloader):
     n_rows = len(downloader.manifest)
     with patch(
-        "datasetinsights.data.simulation.download.download_file"
+        "datasetinsights.datasets.simulation.download.download_file"
     ) as mocked_dl:
         downloader.download_references()
         downloader.download_captures()
