+++ tests/simulation/test_metrics.py
@@ -4,9 +4,12 @@ import json
 import pandas as pd
 import pytest
 
-from datasetinsights.data.simulation import Metrics
-from datasetinsights.data.simulation.exceptions import DefinitionIDError
-from datasetinsights.data.simulation.tables import SCHEMA_VERSION, glob
+from datasetinsights.datasets.simulation import (
+    SCHEMA_VERSION,
+    DefinitionIDError,
+    Metrics,
+    glob,
+)
 
 
 def test_filter_metrics(mock_data_dir):
